Update dependencies and refactor endpoint structure
- Bump versions of several dependencies in Cargo.lock and Cargo.toml, including aws-sdk-sqs to 1.78.0 and lambda_http to 0.16.0.
- Change project edition from 2021 to 2024 in Cargo.toml.
- Refactor the endpoint implementation by moving it to a new module `basic` and updating the response format to return JSON.
- Remove the old endpoint code from lib.rs.
