+++ Cargo.toml
@@ -1,9 +1,10 @@
 [package]
 name = "kinetics-endpoint-template"
 version = "0.1.0"
-edition = "2021"
+edition = "2024"
 
 [dependencies]
-kinetics-macro = "0.2.1"
-lambda_http = "0.14.0"
-aws-sdk-sqs = "1.67.0"
+kinetics-macro = "0.7.0"
+lambda_http = "0.16.0"
+aws-sdk-sqs = "1.78.0"
+serde_json = "1.0.142"
