+++ src/basic/endpoint.rs
@@ -1,0 +1,24 @@
+use aws_sdk_sqs::operation::send_message::builders::SendMessageFluentBuilder;
+use kinetics_macro::endpoint;
+use lambda_http::{Body, Error, Request, Response};
+use serde_json::json;
+use std::collections::HashMap;
+
+/// REST API endpoint which responds with JSON {"success": true}
+///
+/// Test locally with the following command:
+/// kinetics invoke BasicEndpointEndpoint
+#[endpoint(url_path = "/endpoint")]
+pub async fn endpoint(
+    _event: Request,
+    _secrets: &HashMap<String, String>,
+    _queues: &HashMap<String, SendMessageFluentBuilder>,
+) -> Result<Response<Body>, Error> {
+    let resp = Response::builder()
+        .status(200)
+        .header("content-type", "application/json")
+        .body(json!({"success": true}).to_string().into())
+        .map_err(Box::new)?;
+
+    Ok(resp)
+}
